Fix misleading comment in RegistryScript.rgs.
The information stated in the comment was correct at one point, but I
had forgotten to change it when I rearchitectured it.

Previously, the file was generated by a separate file called prebuild.cmd
which was a project-wide pre-build action, but now it uses a MSBuild
custom build tool, which is a cleaner solution.

Signed-off-by: Isabella <kawapure@gmail.com>