Remove addressbarPS (proxy/stub)...
This is an ancient leftover from the ATL project wizard. It even fossilised
the original project name: "addressbar" (instead of "ClassicExplorer").

The proxy/stub project did not build at all at the time of removal.

It seems that its responsibilities can be handled in the main COM server,
so there's no explicit need for a separate project.

For more information, you can read about this feature in this Stack
Overflow answer:
https://stackoverflow.com/a/9584168


Signed-off-by: Isabella <kawapure@gmail.com>