Check in new theme system changes
The new theme system will rely on all themes (except the default XP theme,
which may be subject to change in the future but I plan to keep it as XP
for now) using different resource DLLs for metadata.

This will make customisation of ClassicExplorer a lot easier, however it
requires significantly reworking the present code.

As this is a long-term project, I decided that it's worth publishing into
a separate branch for people to check in on. When I feel that the code is
ready for shipping, it'll be merged back into the main branch and the old
method will be completely abandoned.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>